LayoutTests/fast/forms/ios/select-change-focus-to-another-select-expected.txt

This test verifies that focusing one select and then focusing another does not dismiss the second select's menu.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS areArraysEqual(items, ["D", "E", "F"]) is true
PASS successfullyParsed is true

TEST COMPLETE
